+++ silnlp/nmt/experiment.py
@@ -7,10 +7,10 @@ from typing import Optional, Set
 import yaml
 
 from ..common.environment import SIL_NLP_ENV
+from ..common.postprocesser import PostprocessHandler
 from ..common.utils import get_git_revision_hash, show_attrs
 from .clearml_connection import SILClearML
 from .config import Config, get_mt_exp_dir
-from .postprocess import PostprocessHandler
 from .test import _SUPPORTED_SCORERS, test
 from .translate import TranslationTask
